Moved usermanagement to admin folder
